Split out test suite
